requirements.txt
torch
torchvision
Pillow
numpy
matplotlib 